Restructured the ruby code-base into the respective folders, that correspond to the challenge id per implementation and test for each of the challenge
