+++ lib/solutions/CHK/checkout.rb
