+++ lib/solutions/FIZ/fizz_buzz.rb
