+++ lib/solutions/SUM/sum.rb
