+++ test/SUM/sum_test.rb
@@ -1,5 +1,5 @@
 # noinspection RubyResolve,RubyResolve
-require_relative 'test_helper'
+require_relative '../test_helper'
 require 'logging'
 
 Logging.logger.root.appenders = Logging.appenders.stdout
